4ws driving towards goal
git-svn-id: https://rarepos.cs.uni-tuebingen.de/svn-path/rabot/trunk/ros/navigation@3223 690284e3-2886-4bbc-99f2-a987503dca67

+++ path_following/src/DualPidCtrl.h
@@ -14,7 +14,7 @@ public:
       @param deltaf result steer angle front in rad
       @param deltar result steer angle rear in rad
       */
-    bool execute (double ef, double er, float& deltaf, float& deltar);
+    bool execute (double ef, double er, double& deltaf, double& deltar);
 
     /**
       set sampling time (abtastzeit)
